fix(ci): run spotless before the version updates are committed
The release plugin updates the versions in our pom files. This update will break the formatting that we have configured in spotless. As a result, when we'd like to merge the release branch back to the main branch, we manually need to run spotless:apply in order to restore the formatting.

We can add a completion goal to the release plugin. This parameter will:

"Goals to run on completion of the preparation step, after transformation back to the next development version but before committing. Space delimited."

This means it will execute spotless:apply once the release plugin has updated the versions to the next development version, but before the commit is made. Making the correct format part of the release commit.